update app Statistiche da correggere

+++ Esercitazioni/Gestionale-Personale-mvc/Statistiche.cs
@@ -2,7 +2,7 @@ public class Statistiche{
    // public int Id{ get; set; }
     public int Performance{ get; set; }
 
-    public double Assenze{ get; set; }
+    public int Assenze{ get; set; }
 
     public Statistiche(int performance,int assenze){
